requirements.txt
fastapi==0.95
pydantic==1.10
alembic==1.10
pytest==7.3
sqlalchemy==2.0
uvicorn==0.21.1
python-dotenv==1.0.0
asyncpg
greenlet
httpx
pytest-asyncio
anyio
pytest-tornasync
pytest-trio
pytest-twisted
twisted
gunicorn
psycopg2